# [8_9] Perf tuning on descends