--- a/System/Classes/url.cpp
+++ b/System/Classes/url.cpp
@@ -542,12 +542,12 @@ factor (url u) {
 
 bool
 descends (url u, url base) {
-  if (is_or (u)) return descends (u[1], base) && descends (u[2], base);
-  if (is_or (base)) return descends (u, base[1]) || descends (u, base[2]);
   if (u == base) return true;
-  if (is_concat (u) && is_atomic (base)) return u[1] == base;
   if (is_concat (u) && is_concat (base))
     return u[1] == base[1] && descends (u[2], base[2]);
+  if (is_concat (u) && is_atomic (base)) return u[1] == base;
+  if (is_or (u)) return descends (u[1], base) && descends (u[2], base);
+  if (is_or (base)) return descends (u, base[1]) || descends (u, base[2]);
   return false;
 }