--- a/bench/System/Classes/url_bench.cpp
+++ b/bench/System/Classes/url_bench.cpp
@@ -16,6 +16,9 @@ main () {
   lolly::init_tbox ();
   bench.run ("url descends equal",
              [&] { descends (url ("a/b/c"), url ("a/b/c")); });
-  bench.run ("url descends concat",
-             [&] { descends (url ("a/b/c"), url ("a/b/c/d/e/f")); });
+  bench.run ("url descends concat 2",
+             [&] { descends (url ("a/b/c/d/e/f"), url ("a/b")); });
+  bench.run ("url descends concat 8", [&] {
+    descends (url ("a/b/c/d/e/f/g/h/i/j"), url ("a/b/c/d/e/f/g/h"));
+  });
 }