## Optimization 2
```
# Before
|               ns/op |                op/s |    err% |     total | benchmark
|--------------------:|--------------------:|--------:|----------:|:----------
|            8,400.00 |          119,047.62 |    0.2% |      0.01 | `url descends equal`
|           12,203.41 |           81,944.32 |    0.8% |      0.01 | `url descends concat 2`
|           37,293.00 |           26,814.68 |    0.7% |      0.01 | `url descends concat 8`

# After

|               ns/op |                op/s |    err% |     total | benchmark
|--------------------:|--------------------:|--------:|----------:|:----------
|            8,395.67 |          119,109.03 |    0.2% |      0.01 | `url descends equal`
|           12,189.40 |           82,038.51 |    0.5% |      0.01 | `url descends concat 2`
|           32,826.84 |           30,462.87 |    0.2% |      0.01 | `url descends concat 8`
```

## Optimization 1
Adjust the order (url_or should be put at the end)
```
# Before

|               ns/op |                op/s |    err% |     total | benchmark
|--------------------:|--------------------:|--------:|----------:|:----------
|            8,482.58 |          117,888.63 |    0.3% |      0.01 | `url descends equal`
|           14,931.67 |           66,971.76 |    0.3% |      0.01 | `url descends concat`

# After

|               ns/op |                op/s |    err% |     total | benchmark
|--------------------:|--------------------:|--------:|----------:|:----------
|            8,385.20 |          119,257.80 |    0.1% |      0.01 | `url descends equal`
|           14,816.56 |           67,492.06 |    1.3% |      0.01 | `url descends concat`
```